Move BlockValidator to separate file
